# contracts/openapi_stub.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WAYMARK API",
    "description": "Explainable road risk, cell by cell, plus the action loop: region reports, an assistant, tracked measures with photo evidence and a verified-only adjusted risk estimate. Decision support only; human experts decide.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/measures/{measure_id}/evidence": {
      "post": {
        "tags": [
          "evidence"
        ],
        "summary": "Upload Evidence",
        "operationId": "upload_evidence_api_measures__measure_id__evidence_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "measure_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Measure Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_evidence_api_measures__measure_id__evidence_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Evidence"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evidence/{evidence_id}/file": {
      "get": {
        "tags": [
          "evidence"
        ],
        "summary": "Evidence File",
        "operationId": "evidence_file_api_evidence__evidence_id__file_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "evidence_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Evidence Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evidence/{evidence_id}/thumb": {
      "get": {
        "tags": [
          "evidence"
        ],
        "summary": "Evidence Thumb",
        "operationId": "evidence_thumb_api_evidence__evidence_id__thumb_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "evidence_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Evidence Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_evidence_api_measures__measure_id__evidence_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "kind": {
            "$ref": "#/components/schemas/EvidenceKind"
          },
          "caption": {
            "type": "string",
            "title": "Caption",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "file",
          "kind"
        ],
        "title": "Body_upload_evidence_api_measures__measure_id__evidence_post"
      },
...
      "Evidence": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "measure_id": {
            "type": "integer",
            "title": "Measure Id"
          },
          "kind": {
            "$ref": "#/components/schemas/EvidenceKind"
          },
          "file_url": {
            "type": "string",
            "title": "File Url"
          },
          "thumb_url": {
            "type": "string",
            "title": "Thumb Url"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "geo_flag": {
            "type": "string",
            "enum": [
              "ok",
              "far",
              "missing"
            ],
            "title": "Geo Flag"
          },
          "stale_photo": {
            "type": "boolean",
            "title": "Stale Photo"
          },
          "exif_taken_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exif Taken At"
          },
          "uploaded_by": {
            "type": "string",
            "title": "Uploaded By"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "measure_id",
          "kind",
          "file_url",
          "thumb_url",
          "caption",
          "geo_flag",
          "stale_photo",
          "exif_taken_at",
          "uploaded_by",
          "created_at"
        ],
        "title": "Evidence"
      },
      "EvidenceKind": {
        "type": "string",
        "enum": [
          "before",
          "during",
          "after"
        ],
        "title": "EvidenceKind"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}